requirements.txt
Flask
Flask-SocketIO
python-dotenv
eventlet
Flask-Session